implement auto-adding the default role finally
